+++ src/fft_component.hpp
@@ -41,10 +41,10 @@ namespace audio_analysis_lib
 
 	DLLEXPORT fft_component::fft_exception fft_mode_setting(fft_component::fft_mode mode, void* func_object);
 
-	DLLEXPORT int get_fft_size(void* func_object);
+	DLLEXPORT int32_t get_fft_size(void* func_object);
 
 	DLLEXPORT void delete_fft_component(void** func_object);
 
-	DLLEXPORT size_t get_fft_component_size();
+	DLLEXPORT int32_t get_fft_component_size();
 }
